% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NShjelpefunksjoner.R
\name{abonnement}
\alias{abonnement}
\title{Funksjon som produserer rapporten som skal sendes til mottager.
(The actual call to this function is made through do.call and 
has the effect of providing the parameters as class
\emph{list}. Verdier gis inn som listeparametre}
\usage{
abonnement(rnwFil, brukernavn = "tullebukk", reshID = 0,
  datoFra = Sys.Date() - 400, datoTil = Sys.Date())
}
\arguments{
\item{reshID}{Aktuell reshid}

\item{datoFra}{dato}

\item{parametre}{Liste med valgfrie parametre, avhengig av type rapport}
}
\value{
Full path of file produced
}
\description{
Funksjon som produserer rapporten som skal sendes til mottager.
(The actual call to this function is made through do.call and 
has the effect of providing the parameters as class
\emph{list}. Verdier gis inn som listeparametre
}
