% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NShjelpefunksjoner.R
\name{finnRegData}
\alias{finnRegData}
\title{Hente data basert på valgtVar}
\usage{
finnRegData(valgtVar = "Alder", Data = AlleTab)
}
\arguments{
\item{valgtVar}{Angir hvilke(n) variable det skal vises resultat for.}

\item{Data}{Liste med alle skjema/tabeller}
}
\description{
Hente data basert på valgtVar
}
