% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\alias{getRealData}
\alias{getFakeData}
\alias{processData}
\title{Get Data for NordicScir}
\usage{
getRealData(register = "norscir")

getFakeData()

processData(register = "norscir", data)
}
\arguments{
\item{register}{Hvilket register det skal hentes data for}

\item{data}{A data object to be processed by the function.}
}
\value{
A list of data objects (data frames).  If not successful in query for
real data, a warning is issued and NULL will be will be returned.
}
\description{
Functions to obtain and process data for NordicScir, both real and fake.
}
