% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NShjelpefunksjoner.R
\name{kjor_NSapper}
\alias{kjor_NSapper}
\title{Kjør Shiny Application}
\usage{
kjor_NSapper(register = "norscir")
}
\arguments{
\item{register}{Angi hvilket register som skal startes}
}
\value{
Et objekt som representerer den aktuelle app'en
}
\description{
Kjør Shiny Application
}
