% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NShjelpefunksjoner.R
\name{KobleMedHoved}
\alias{KobleMedHoved}
\title{Koble med aktuell hovedtabell}
\usage{
KobleMedHoved(HovedSkjema, Skjema2, alleHovedskjema = F, alleSkjema2 = F)
}
\arguments{
\item{HovedSkjema}{hvilket skjema skal oppfølgingsskjemaet kobles til? De fleste kobles til Hovedskjema. 
Satisfaction kobles til Performance}

\item{Skjema2}{Oppfølgingsskjemaet som skal kobles til sitt aktuelle hovedskjema}

\item{alleHovedskjema}{TRUE (FALSE) Om alle registreringer fra hovedskjemaet skal være med (TRUE), eller bare
de registreringer som har ei oppfølgning (FALSE).}
}
\description{
Koble med aktuell hovedtabell
}
