% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NShjelpefunksjoner.R
\name{lageTulleData}
\alias{lageTulleData}
\title{Lage tulledata (simulerte data). Takler ikke posixlt- type data.}
\usage{
lageTulleData(RegData, varBort = "ShNavn", antSh = 27)
}
\description{
Lage tulledata (simulerte data). Takler ikke posixlt- type data.
}
