% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NStabeller.R
\name{lagTabavFigAndeler}
\alias{lagTabavFigAndeler}
\title{Vise figurdata som tabell, fordelingsfigur}
\usage{
lagTabavFigAndeler(UtDataFraFig)
}
\arguments{
\item{UtDataFraFig}{Dataliste fra figuren}
}
\description{
Vise figurdata som tabell, fordelingsfigur
}
