% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NStabeller.R
\name{lagTabavFigAndelerSh}
\alias{lagTabavFigAndelerSh}
\title{Vise figurdata som tabell, fordeling per sykehus}
\usage{
lagTabavFigAndelerSh(UtDataFraFig)
}
\arguments{
\item{UtDataFraFig}{Dataliste fra figuren}
}
\description{
Vise figurdata som tabell, fordeling per sykehus
}
