% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NStabeller.R
\name{lagTabNevrKlass}
\alias{lagTabNevrKlass}
\title{Nevrologisk klassifikasjon}
\usage{
lagTabNevrKlass(
  HovedSkjema,
  datoFra = "2018-01-01",
  datoTil = Sys.Date(),
  datoUt = 0
)
}
\arguments{
\item{datoFra}{startdato, fra og med Eks. '2010-01-01'    # min og max dato i utvalget vises alltid i figuren.}

\item{datoTil}{sluttdato, fra og med Eks. '2013-05-25'}

\item{datoUt}{datofiltrering basert på innleggelse (0), utskriving (1)}

\item{HovedSkjema-}{hovedskjema}
}
\description{
Nevrologisk klassifikasjon
}
