% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSFigAndelerGrVar.R
\name{NSFigAndelerGrVar}
\alias{NSFigAndelerGrVar}
\title{Søylediagram med andeler for hver grupperingsenhet (sykehus, RHF, ...)}
\usage{
NSFigAndelerGrVar(
  RegData = 0,
  preprosess = 1,
  valgtVar = "ABMI",
  datoFra = "2015-01-01",
  datoTil = Sys.Date(),
  minald = 0,
  maxald = 130,
  erMann = "",
  enhetsUtvalg = 0,
  grVar = "ShNavn",
  Ngrense = 10,
  reshID = 0,
  outfile = "",
  ...
)
}
\arguments{
\item{RegData}{- ei dataramme med alle nødvendige variable fra registeret.}

\item{preprosess}{Preprosesser data
0: Nei
1: Ja (Standard)}

\item{valgtVar}{- variabelen det skal genereres resultat for}

\item{datoFra}{startdato, fra og med Eks. '2010-01-01'    # min og max dato i utvalget vises alltid i figuren.}

\item{datoTil}{sluttdato, fra og med Eks. '2013-05-25'}

\item{minald}{- alder, fra og med}

\item{maxald}{- alder, til og med}

\item{erMann}{- kjønn, 1-menn, 0-kvinner, standard: '' (alt annet enn 0 og 1), dvs. begge}

\item{enhetsUtvalg}{- 0:hele landet/Norden (standard),
1:eget sykehus sml med resten av landet/Norden,
2:eget sykehus
3:eget sykehus mot eget land
4:eget land
5:eget land mot andre land}

\item{grVar}{Tekstvariabel som angir hva skal resultatene grupperes på.
ShNavn-sykehus/avdeling
Fylke- Pasienten bor i det akutelle fylket}

\item{Ngrense}{Minste antall registreringer for at ei gruppe skal bli vist}

\item{reshID}{- avdelingsid for egen avdeling, standard: 0-hele landet.}

\item{outfile}{- navn på fil figuren skrives ned til.}
}
\value{
Figur med...
}
\description{
Funksjon som genererer en figur med andeler av en variabel for en valgt grupperingsvariabel,
f.eks. sykehus.
Funksjonen er delvis skrevet for å kunne brukes til andre grupperingsvariable enn sykehus
}
\details{
Andel som mottar sykepenger er definert som svaralternativene: 'Sykemeldt',
       'Aktiv sykemeldt', 'Delvis sykemeldt', 'Attføring/rehab.', 'Uføretrygdet'
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item ABMI: Målt og veid ved innleggelse?
    \item FBMI: Målt og veid ved utskriving?
	}
}
