% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSFigAndelStabel.R
\name{NSFigAndelStabel}
\alias{NSFigAndelStabel}
\title{Søylediagram som sammenligner fordeling (vha. stablede søyler) av valgt variabel}
\usage{
NSFigAndelStabel(
  RegData,
  outfile = "",
  valgtVar,
  datoFra = "2010-01-01",
  datoTil = "3000-01-01",
  minald = 0,
  maxald = 130,
  erMann = "",
  traume = 99,
  preprosess = 1,
  enhetsUtvalg = enhetsUtvalg,
  reshID,
  hentData = 0,
  ...
)
}
\arguments{
\item{RegData}{- ei dataramme med alle nødvendige variable fra registeret.}

\item{outfile}{- navn på fil figuren skrives ned til.}

\item{valgtVar}{- variabelen det skal genereres resultat for}

\item{datoFra}{<- '2010-01-01'    # min og max dato i utvalget vises alltid i figuren.}

\item{datoTil}{<- '2013-05-25'}

\item{minald}{- alder, fra og med}

\item{maxald}{- alder, til og med}

\item{erMann}{- kjønn, 1-menn, 0-kvinner, standard: '' (alt annet enn 0 og 1), dvs. begge}

\item{traume}{- 'ja','nei', 'alle' standard: ikke valgt}

\item{preprosess}{Preprosesser data
0: Nei
1: Ja (Standard)}

\item{enhetsUtvalg}{- 1:eget sykehus, 0:hele landet (standard) Kun for valgtVar=='NevrNivaaInnUt'}

\item{reshID}{- avdelingsid for egen avdeling, standard: 0-hele landet.}

\item{hentData}{Gjør spørring mot database hvis data ikke er levert fra andre kilder.
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}
}
\description{
Søylediagram som sammenligner fordeling (vha. stablede søyler) av valgt variabel
hos fra to ulike utvalg, f.eks. eget/resten, før/etter
(Kan velge å ikke ha med sammenligninga)
Søylediagram som viser andeler av valgt variabel:
NevrNivaaInn' - Nevrologisk nivå ved innskrivelse, sml. eget, andre
NevrNivaaUt' - Nevrologisk nivå ved utskrivelse, sml. eget, andre
NevrNivaaInnUt' - sml. Nevrologisk nivå ved inn- og utskrivelse
Hvilke kombinasjonsmuligheter har vi?
Inn / Ut (Velge: bare eget/hele landet)
Eget / resten av landet (Velge: Inn eller Ut)
}
