% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSFigGjsnGrVar.R
\name{NSFigGjsnGrVar}
\alias{NSFigGjsnGrVar}
\title{Funksjon som genererer en figur med gjennomsnitt/median}
\usage{
NSFigGjsnGrVar(
  RegData,
  valgtVar,
  valgtMaal = "gjsn",
  grVar = "ShNavn",
  datoFra = "2010-01-01",
  datoTil = "2050-12-31",
  datoUt = 0,
  AIS = "",
  minald = 0,
  maxald = 130,
  erMann = "",
  traume = "alle",
  nivaaUt = 99,
  preprosess = 1,
  outfile = "",
  hentData = 0,
  ...
)
}
\arguments{
\item{RegData}{- ei dataramme med alle nødvendige variable fra registeret.}

\item{valgtVar}{- variabelen det skal genereres resultat for}

\item{valgtMaal}{- 'med' = median. Alt annet gir gjennomsnitt}

\item{datoFra}{startdato, fra og med Eks. '2010-01-01'    # min og max dato i utvalget vises alltid i figuren.}

\item{datoTil}{sluttdato, fra og med Eks. '2013-05-25'}

\item{datoUt}{Om man skal velge inn eller ut-dato som grunnlag for datofiltrering}

\item{AIS}{- AISgrad ved innleggelse alle(''), velge en eller flere fra 1:5, mappes til: A,B,C,D,E}

\item{minald}{- alder, fra og med}

\item{maxald}{- alder, til og med}

\item{erMann}{- kjønn, 1-menn, 0-kvinner, standard: '' (alt annet enn 0 og 1), dvs. begge}

\item{traume}{- 'ja','nei', 'alle' standard: ikke valgt}

\item{nivaaUt}{- Nivå ved utreise, flervalgs 0:tetraplegi, 1:paraplegi, 2:C1-4, 3:C5-8, 9:ukjent}

\item{preprosess}{Preprosesser data
0: Nei
1: Ja (Standard)}

\item{outfile}{- navn på fil figuren skrives ned til.}

\item{hentData}{Gjør spørring mot database hvis data ikke er levert fra andre kilder.
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}
}
\description{
Funksjon som genererer en figur med gjennomsnitt/median for hvert sykehus
og kan ta inn ulike numeriske variable. Funksjonen er delvis skrevet for å
kunne brukes til andre grupperingsvariable enn sykehus.
}
\details{
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item Alder: Aldersfordeling, 15-årige grupper 
    \item DagerRehab: antall dager med rehabilitering [RehabDy]
    \item DagerTilRehab: antall dager før rehabilitering [BeforeRehDy]
    \item OpphTot: Lengde på opphold – totalt [HosptlDy] 
    \item Permisjon: Antall døgn ute av sykehus [OutOfHosptlDy]
   }
}
