% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSFigGjsnTid.R
\name{NSFigGjsnTid}
\alias{NSFigGjsnTid}
\title{Utvikling over tid for gjennomsnitt/median av valgt variabel}
\usage{
NSFigGjsnTid(
  RegData,
  valgtVar = "Alder",
  datoFra = "2011-01-01",
  datoTil = "3000-12-31",
  tidsenhet = "Aar",
  minald = 0,
  maxald = 110,
  erMann = "",
  reshID = 0,
  enhetsUtvalg = 0,
  valgtMaal = "Gjsn",
  AIS = "",
  traume = "alle",
  nivaaUt = 99,
  preprosess = 1,
  outfile = "",
  hentData = 0,
  lagFigur = 1,
  ...
)
}
\arguments{
\item{RegData}{- ei dataramme med alle nødvendige variable fra registeret.}

\item{valgtVar}{Hvilken variabel som skal visualiseres}

\item{reshID}{- avdelingsid for egen avdeling, standard: 0-hele landet.}

\item{valgtMaal}{'gjsn': gir middelverdi (standard)
'med': gir median}

\item{preprosess}{Preprosesser data
0: Nei
1: Ja (Standard)}

\item{outfile}{- navn på fil figuren skrives ned til.}

\item{hentData}{Gjør spørring mot database hvis data ikke er levert fra andre kilder.
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}
}
\value{
Linjediagram som viser utvikling over tid for valgt variabel
}
\description{
Figuren viser gjennomsnitt/median per år med konfidensintervall for valgt variabel.
I bakgrunn vises konfidensintervall for det man har valt å sammenlikne med.
}
\details{
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item alder: Pasientens alders
   }
}
