% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSFigKvalIndIkkeNodvendigIkkeKlar.R
\name{NSFigKvalInd}
\alias{NSFigKvalInd}
\title{Søylediagram som viser flere kvalitetsindikatorer}
\usage{
NSFigKvalInd(
  NSdata,
  outfile = "",
  valgtVar,
  datoFra = "2010-01-01",
  datoTil = "2050-01-01",
  AIS = "",
  minald = 0,
  maxald = 130,
  erMann = 99,
  traume = "",
  nivaaUt = 99,
  enhetsUtvalg = 1,
  reshID,
  hentData = 0,
  preprosess = 1
)
}
\arguments{
\item{NSdata}{- Alle nødvendige variable fra hvert skjema.}

\item{outfile}{- navn på fil figuren skrives ned til.}

\item{valgtVar}{- variabelen det skal genereres resultat for}

\item{datoFra}{<- '2010-01-01'    # min og max dato i utvalget vises alltid i figuren.}

\item{datoTil}{<- '2013-05-25'}

\item{AIS}{- AISgrad ved innleggelse alle(''), velge en eller flere fra 1:5, mappes til: A,B,C,D,E}

\item{minald}{- alder, fra og med}

\item{maxald}{- alder, til og med}

\item{erMann}{- kjønn, 1-menn, 0-kvinner, standard: '' (alt annet enn 0 og 1), dvs. begge}

\item{traume}{- 'ja','nei', 'alle' standard: ikke valgt}

\item{nivaaUt}{- Nivå ved utreise, flervalgs tetraplegi, paraplegi, C1-4, C5-8, ukjent}

\item{enhetsUtvalg}{- 1:eget sykehus sml med resten, 2:eget sykehus, 0:hele landet (standard)}

\item{reshID}{- avdelingsid for egen avdeling, standard: 0-hele landet.}

\item{hentData}{Gjør spørring mot database hvis data ikke er levert fra andre kilder.
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{preprosess}{Preprosesser data
0: Nei
1: Ja (Standard)}

\item{figurtype}{angir hvilken figurtype som skal lages: andeler, gjsnGrVar}
}
\value{
Søylediagram samling av kvalitetsindikatorer
}
\description{
Funksjonen genererer en figur med ei gruppe kvalitetsindikatorer.
}
\details{
Argumentet \emph{valgtVar} kan bare være "kvalInd".
}
