% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSFigPrePost.R
\name{NSFigPrePost}
\alias{NSFigPrePost}
\title{Søylediagram som viser resultat av valgt variabel, målt ved to tidspunkter}
\usage{
NSFigPrePost(
  RegData,
  valgtVar = "KontUtTil",
  datoFra = "2019-01-01",
  datoTil = Sys.Date(),
  enhetsUtvalg = 0,
  datoUt = 0,
  reshID = 0,
  velgAvd = 0,
  minald = 0,
  maxald = 130,
  register = "norscir",
  erMann = 9,
  traume = "alle",
  AIS = "",
  nivaaUt = 99,
  Ngrense = 10,
  outfile = "",
  preprosess = 0,
  hentData = 0,
  ...
)
}
\arguments{
\item{RegData}{er her KontrollH, dvs. Kontrollskjema koblet til hovedskjema}

\item{valgtVar}{Hvilken variabel som skal visualiseres}

\item{enhetsUtvalg}{foreløpig 0:hele landet/Norden (standard),
2:eget sykehus
4:eget land}

\item{datoUt}{Om man skal velge inn eller ut-dato som grunnlag for datofiltrering}

\item{reshID}{- avdelingsid for egen avdeling, standard: 0-hele landet.}

\item{outfile}{- navn på fil figuren skrives ned til.}

\item{preprosess}{Preprosesser data
0: Nei
1: Ja (Standard)}

\item{hentData}{Gjør spørring mot database hvis data ikke er levert fra andre kilder.
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}
}
\value{
Søylediagram som fordeling av valgt variabel, ved operasjon, samt 1. kontroll
}
\description{
Funksjon som genererer en figur med som viser endring i en variabels fordeling ved to ulike tidspunkter.
}
\details{
Detajer: Her bør man liste opp hvilke variable funksjonen benytter...
}
