% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSUtvalg.R
\name{NSUtvalg}
\alias{NSUtvalg}
\title{Funksjon som gjør utvalg av data basert på brukervalg.}
\usage{
NSUtvalg(
  RegData,
  datoFra = "2010-01-01",
  datoTil = Sys.Date(),
  minald = 0,
  maxald = 110,
  erMann = 99,
  traume = "alle",
  AIS = "",
  enhetsUtvalg = 0,
  nivaaUt = 99,
  reshID = 0,
  fargepalett = "BlaaOff",
  datoUt = 0
)
}
\arguments{
\item{datoFra}{<- '2010-01-01'    # min og max dato i utvalget vises alltid i figuren.}

\item{datoTil}{<- '2013-05-25'}

\item{minald}{- alder, fra og med}

\item{maxald}{- alder, til og med}

\item{erMann}{- kjønn, 1-menn, 0-kvinner, standard: '' (alt annet enn 0 og 1), dvs. begge}

\item{traume}{- 'ja','nei', 'alle' standard: ikke valgt}

\item{AIS}{- AISgrad ved innleggelse alle(''), velge en eller flere fra 1:5, mappes til: A,B,C,D,E}

\item{enhetsUtvalg}{- 1:eget sykehus sml med resten, 2:eget sykehus, 0:hele landet (standard)}

\item{nivaaUt}{- Nivå ved utreise, flervalgs tetraplegi, paraplegi, C1-4, C5-8, ukjent}
}
\description{
Funksjon som gjør utvalg av dataene, returnerer det reduserte datasettet og
utvalgsteksten.
}
