% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSUtvalgEnh.R
\name{NSUtvalgEnh}
\alias{NSUtvalgEnh}
\title{Funksjon som gjør utvalg av data basert på brukervalg.}
\usage{
NSUtvalgEnh(
  RegData,
  datoFra = "2010-01-01",
  datoTil = Sys.Date(),
  datoUt = 0,
  minald = 0,
  maxald = 110,
  erMann = 99,
  traume = "alle",
  AIS = "",
  enhetsUtvalg = 0,
  nivaaUt = 99,
  reshID = 0,
  fargepalett = "BlaaOff"
)
}
\arguments{
\item{datoFra}{startdato, fra og med Eks. '2010-01-01'    # min og max dato i utvalget vises alltid i figuren.}

\item{datoTil}{sluttdato, fra og med Eks. '2013-05-25'}

\item{datoUt}{datofiltrering basert på innleggelse (0), utskriving (1)}

\item{minald}{- alder, fra og med}

\item{maxald}{- alder, til og med}

\item{erMann}{- kjønn, 1-menn, 0-kvinner, standard: '' (alt annet enn 0 og 1), dvs. begge}

\item{traume}{- 'ja','nei', 'alle' standard: ikke valgt}

\item{AIS}{- AISgrad ved innleggelse alle(''), velge en eller flere fra 1:5, mappes til: A,B,C,D,E}

\item{enhetsUtvalg}{- 0:hele landet/Norden (standard),
1:eget sykehus sml med resten av landet/Norden,
2:eget sykehus
3:eget sykehus mot eget land
4:eget land
5:eget land mot andre land}

\item{nivaaUt}{- Nivå ved utreise, flervalgs 0:tetraplegi, 1:paraplegi, 2:C1-4, 3:C5-8, 9:ukjent}
}
\description{
Funksjon som gjør utvalg av dataene, returnerer det reduserte datasettet og
utvalgsteksten. Funskjonen genererer også utvalg (indekser) for sammenligning av enhetsgrupper.
Tekster tilpasses nordisk register.
}
