% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSVarTilrettelegg.R
\name{NSVarTilrettelegg}
\alias{NSVarTilrettelegg}
\title{Funksjon for å tilrettelegge variable for beregning.}
\usage{
NSVarTilrettelegg(RegData, valgtVar, grVar = "", figurtype = "andeler")
}
\arguments{
\item{RegData}{- ei dataramme med alle nødvendige variable fra registeret.}

\item{valgtVar}{- variabelen det skal genereres resultat for}

\item{figurtype}{angir hvilken figurtype som skal lages: andeler, gjsnGrVar}
}
\value{
Definisjon av valgt variabel.
}
\description{
Denne funksjonen gjør utvalg og tilrettelegger variable (gitt ved valgtVar) til videre bruk. 
Videre bruk kan eksempelvis være beregning av AggVerdier eller gjennomsnitt. 
Funksjonen gjør også filtreringer som å fjerne ugyldige verdier for den valgte variabelen, samt ta høyde for avhengigheter med
andre variable. Det er også her man angir aksetekster og titler for den valgte variabelen. 
Her kan mye hentes til analysebok
}
\details{
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item alder: Aldersfordeling, 10-årige grupper 
   }
   
NB: valgtVar må komponeres slik at den henspeiler på hvilken 
tabell variabelen kommer fra. Dette for å senere kunne hente og koble dataene riktig.
valgtSkjema <- substr(valgtVar,1,4) 
Prefixet til  \emph{valgtVar} skal være følgende:
   \itemize{
   \item Livs - LifeQuality, 
   \item Urin - UrinaryTractFunction
   \item Tarm - BowelFunction
   \item Kont - Control
   \item Funk - ActivityAndParticipationPerformance
   \item Tilf - ActivityAndParticipationSatisfaction
        }
Argumentet \emph{enhetsUtvalg} har følgende valgmuligheter:
   \itemize{
    \item 0: Hele landet
    \item 1: Egen enhet mot resten av landet (Standard)
    \item 2: Egen enhet
   	}
}
