% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_nordicscir.R
\name{server_nordicscir}
\alias{server_nordicscir}
\title{Server-del til appen}
\usage{
server_nordicscir(input, output, session)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}
}
\value{
Server-delen til nordicscir-appen
}
\description{
Server-del til appen
}
