% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NStabeller.R
\name{tabAntOpphPasSh5Aar}
\alias{tabAntOpphPasSh5Aar}
\title{Antall registreringer/pasienter siste 5 år}
\usage{
tabAntOpphPasSh5Aar(RegData, gr = "opph", datoTil)
}
\arguments{
\item{gr}{gruppering 'opph'-opphold, 'pas'-pasient}
}
\description{
Antall registreringer/pasienter siste 5 år
}
