% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NStabeller.R
\name{tabAntOpphSh5Aar}
\alias{tabAntOpphSh5Aar}
\title{Antall opphold siste 5 år}
\usage{
tabAntOpphSh5Aar(RegData, datoTil = Sys.Date(), traume = "")
}
\arguments{
\item{datoTil}{<- '2013-05-25'}

\item{traume}{- 'ja','nei', 'alle' standard: ikke valgt}
}
\description{
Antall opphold siste 5 år
}
