% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NStabeller.R
\name{tabAntOpphShAar}
\alias{tabAntOpphShAar}
\title{Antall opphold siste år}
\usage{
tabAntOpphShAar(RegData, datoTil = Sys.Date(), antAar = 10, traume = "")
}
\arguments{
\item{datoTil}{sluttdato, fra og med Eks. '2013-05-25'}

\item{traume}{- 'ja','nei', 'alle' standard: ikke valgt}
}
\description{
Antall opphold siste år
}
