% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NStabeller.R
\name{tabAntOpphShMnd}
\alias{tabAntOpphShMnd}
\title{tabAntOpphShMnd antall opphold siste X (antMnd) mnd
RegData må inneholde Aar.}
\usage{
tabAntOpphShMnd(RegData, datoTil = Sys.Date(), traume = "", antMnd = 12)
}
\arguments{
\item{datoTil}{sluttdato, fra og med Eks. '2013-05-25'}

\item{traume}{- 'ja','nei', 'alle' standard: ikke valgt}
}
\description{
tabAntOpphShMnd antall opphold siste X (antMnd) mnd
RegData må inneholde Aar.
}
