% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NStabeller.R
\name{tabAntOpphShTid}
\alias{tabAntOpphShTid}
\title{Antall opphold siste år/kvartal/måneder}
\usage{
tabAntOpphShTid(
  RegData,
  datoTil = Sys.Date(),
  tidsenhet = "Aar",
  antTidsenh = 10,
  datoUt = 0,
  traume = ""
)
}
\arguments{
\item{datoTil}{sluttdato, fra og med Eks. '2013-05-25'}

\item{datoUt}{datofiltrering basert på innleggelse (0), utskriving (1)}

\item{traume}{- 'ja','nei', 'alle' standard: ikke valgt}
}
\description{
Antall opphold siste år/kvartal/måneder
}
