% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NStabeller.R
\name{tabAntSkjema}
\alias{tabAntSkjema}
\title{Tabell: Antall av ulike typer skjema}
\usage{
tabAntSkjema(Data = AlleTab, datoFra = "2017-01-01",
  datoTil = Sys.Date())
}
\description{
Tabell: Antall av ulike typer skjema
}
