% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NStabeller.R
\name{tabBelegg}
\alias{tabBelegg}
\title{Funksjoner for å lage tabeller Group of functions page title}
\usage{
tabBelegg(
  RegData,
  tidsenhet = "Mnd",
  datoTil = Sys.Date(),
  enhetsUtvalg = 0,
  reshID = 0
)
}
\arguments{
\item{RegData}{data}

\item{datoTil}{til og med dato}

\item{reshID}{- avdelingsid for egen avdeling, standard: 0-hele landet.}

\item{personIDvar}{Variabelen som angir pasientidentifikasjon}

\item{Data}{Liste med alle datatabeller/skjema}

\item{datoFra}{fra og med dato}
}
\description{
Fil som beregner div tabeller.Group of functions Description section
}
\details{
Fil som inneholder funksjoner for å lage tabeller, i første rekke tellinger av personer
Belegg (samlerapport), antall opphold per år og enhet, ant. pasienter per år og enhet, ant opph. per måned og enhet.
}
