% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NStabeller.R
\name{tabLiggetider}
\alias{tabLiggetider}
\title{Liggetider. Liggetid tot.(OpphTot), Liggetid rehab.(DagerRehab), Liggetid før rehab (DagerTilRehab): 
min/max, gjsn, median}
\usage{
tabLiggetider(RegData, datoFra = "2018-01-01", datoTil = Sys.Date(),
  enhetsUtvalg = 0, reshID = 0, traume = "")
}
\arguments{
\item{RegData}{data}

\item{datoFra}{fra og med dato}

\item{datoTil}{til og med dato}

\item{enhetsUtvalg}{- 1:eget sykehus, 0:hele landet (standard)}

\item{traume}{- 'ja','nei', 'alle' standard: ikke valgt}

\item{Data}{Liste med alle datatabeller/skjema}
}
\description{
Liggetider. Liggetid tot.(OpphTot), Liggetid rehab.(DagerRehab), Liggetid før rehab (DagerTilRehab): 
min/max, gjsn, median
}
