% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NShjelpefunksjoner.R
\name{TilLogiskeVar}
\alias{TilLogiskeVar}
\title{Konvertere boolske variable fra tekst til boolske variable...
Boolske variabler kommer fra jun 2025 som 0-1. Funksjonen kan ikke brukes lenger}
\usage{
TilLogiskeVar(Skjema)
}
\arguments{
\item{valgtVar}{Angir hvilke(n) variable det skal vises resultat for.}

\item{Data}{Liste med alle skjema/tabeller}
}
\description{
Konvertere boolske variable fra tekst til boolske variable...
Boolske variabler kommer fra jun 2025 som 0-1. Funksjonen kan ikke brukes lenger
}
