% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_nordicscir.R
\name{ui_nordicscir}
\alias{ui_nordicscir}
\title{Brukergrensesnitt (ui) til nordscir-appen}
\usage{
ui_nordicscir()
}
\value{
Brukergrensesnittet (ui) til nordscir-appen
}
\description{
Brukergrensesnitt (ui) til nordscir-appen
}
