% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NShjelpefunksjoner.R
\name{abonnement}
\alias{abonnement}
\title{Gererere månedsrapport for abonnement}
\usage{
abonnement(
  rnwFil,
  brukernavn = "ukjent",
  reshID = 0,
  register = "nordicscir",
  datoFra = Sys.Date() - 400,
  datoTil = Sys.Date()
)
}
\arguments{
\item{rnwFil}{navn på Rnw-fila som skal kjøres ('mndRapp.Rnw')}

\item{brukernavn}{abonnentens brukernavn ved bestilling}

\item{reshID}{- reshID som abonnenten var logget inn med ved bestilling}

\item{datoFra}{- startdato for data som hentes til bruk i rapporten}

\item{datoTil}{- sluttdato for data som hentes til bruk i rapporten}
}
\description{
Gererere månedsrapport for abonnement
}
