% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NShjelpefunksjoner.R
\name{hjelpeFunksjoner}
\alias{hjelpeFunksjoner}
\alias{SorterOgNavngiTidsEnhet}
\alias{lageTulleData}
\alias{delTekst}
\alias{KobleMedHoved}
\alias{finnRegData}
\alias{TilLogiskeVar}
\title{Hjelpefunksjoner. Group of functions page title}
\usage{
SorterOgNavngiTidsEnhet(RegData, tidsenhet = "Aar")

lageTulleData(RegData, varBort = "ShNavn", antSh = 27)

delTekst(x, len)

KobleMedHoved(HovedSkjema, Skjema2, alleHovedskjema = F, alleSkjema2 = F)

finnRegData(valgtVar = "Alder", Data = AlleTab)

TilLogiskeVar(Skjema)
}
\arguments{
\item{RegData}{data}

\item{x}{En tekststreng eller vektor av tekststrenger}

\item{len}{Lengden strengen skal brytes ved}

\item{HovedSkjema}{hvilket skjema skal oppfølgingsskjemaet kobles til? De fleste kobles til Hovedskjema. 
Satisfaction kobles til Performance}

\item{Skjema2}{Oppfølgingsskjemaet som skal kobles til sitt aktuelle hovedskjema}

\item{alleHovedskjema}{TRUE (FALSE) Om alle registreringer fra hovedskjemaet skal være med (TRUE), eller bare
de registreringer som har ei oppfølgning (FALSE).}

\item{valgtVar}{Angir hvilke(n) variable det skal vises resultat for.}

\item{Data}{Liste med alle skjema/tabeller}

\item{PasientID}{Variabelen som angir pasientidentifikasjon}

\item{valgtVar}{Angir hvilke(n) variable det skal vises resultat for.}

\item{Data}{Liste med alle skjema/tabeller}
}
\value{
Div hjelpefunksjoner
}
\description{
Fil med div hjelpefunksjoner.Group of functions Description section
}
\details{
Detaljer. kommer senereGroup of functions Details paragraph.
}
\section{Finne reinnleggelser After function section}{

Despite its location, this actually comes after the function section.
Fil som inneholder hjelpefunksjoner. 
SorterOgNavngiTidsEnhet Legger til tidsenhetene Aar, Halvaar, Mnd og Kvartal
}

\section{Tilrettelegge tidsenhetvariabel}{

Probably better if all sections come first, uless have one section per function. Makes it easier to
see the information flow.
}

\section{Lage tulledata (simulerte data). Takler ikke posixlt- type data.}{
NA
}

\section{Automatisk linjebryting av lange tekstetiketter}{
NA
}

\section{Koble med aktuell hovedtabell}{
NA
}

\section{Hente data basert på valgtVar}{
NA
}

\section{Konvertere boolske variable fra tekst til boolske variable...}{
NA
}

