% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NStabeller.R
\name{lagTabNevrKlass}
\alias{lagTabNevrKlass}
\title{Nevrologisk klassifikasjon}
\usage{
lagTabNevrKlass(HovedSkjema, datoFra = "2018-01-01",
  datoTil = Sys.Date())
}
\arguments{
\item{datoFra}{<- '2010-01-01'    # min og max dato i utvalget vises alltid i figuren.}

\item{datoTil}{<- '2013-05-25'}

\item{HovedSkjema-}{hovedskjema}
}
\description{
Nevrologisk klassifikasjon
}
