% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NStabeller.R
\name{NordicScirtabeller}
\alias{NordicScirtabeller}
\alias{tabBelegg}
\alias{tabAntOpphShMnd}
\alias{tabAntOpphSh5Aar}
\alias{tabAntOpphPasSh5Aar}
\alias{finnDblReg}
\alias{lagTabavFig}
\title{Funksjoner for å lage tabeller Group of functions page title}
\usage{
tabBelegg(RegData, tidsenhet = "Aar", datoTil, enhetsUtvalg = 0,
  reshID = 0)

tabAntOpphShMnd(RegData, datoTil, antMnd = 6)

tabAntOpphSh5Aar(RegData, datoTil)

tabAntOpphPasSh5Aar(RegData, gr = "opph", datoTil)

finnDblReg(RegData, datoFra = "2017-01-01", datoTil = Sys.Date(),
  reshID = 114240)

lagTabavFig(UtDataFraFig)
}
\arguments{
\item{RegData}{data}

\item{datoTil}{sluttdato. Brukes i tabellene AntOpph per 12 mnd og Belegg}

\item{personIDvar}{Variabelen som angir pasientidentifikasjon}
}
\value{
Div tabeller
}
\description{
Fil som beregner div tabeller.Group of functions Description section
}
\details{
Detaljer. kommer senereGroup of functions Details paragraph.

Fil som inneholder funksjoner for å lage tabeller, i første rekke tellinger av personer 
Aktuelle tabeller:
-Belegg (samlerapport), antall opphold per år og enhet, ant. pasienter per år og enhet, ant opph. per måned og enhet.
-tabAntOpphSh12mnd: Antall opphold per måned og enhet siste 12 måneder fram til datoTil. 
RegData må inneholde InnDato.
-tabAntOpphSh5Aar:Antall opphold per år og enhet siste 5 år (inkl. inneværende år) fram til datoTil. 
RegData må inneholde Aar.
}
\section{Belegg (antall opphold, pasienter og intensivdøgn)}{
NA
}

\section{tabAntOpphShMnd antall opphold siste X (antMnd) mnd}{
NA
}

\section{Antall opphold siste 5 år}{
NA
}

\section{Antall registreringer/pasienter siste 5 år}{

Hmmm
}

\section{Finn eventuelle dobbeltregistreringer}{
NA
}

\section{Vise figurdata som tabell}{
NA
}

