% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NStabeller.R
\name{NordicScirtabeller}
\alias{NordicScirtabeller}
\alias{tabBelegg}
\alias{tabLiggetider}
\alias{tabAntOpphShMnd}
\alias{tabAntOpphSh5Aar}
\alias{tabAntOpphPasSh5Aar}
\alias{tabSkjemaTilknyttet}
\alias{tabAntSkjema}
\alias{lagTabavFigAndeler}
\alias{lagTabavFigGjsnGrVar}
\alias{lagTabNevrKlass}
\title{Funksjoner for å lage tabeller Group of functions page title}
\usage{
tabBelegg(RegData, tidsenhet = "Mnd", datoTil = Sys.Date(),
  enhetsUtvalg = 0, reshID = 0)

tabLiggetider(RegData, datoFra = "2018-01-01", datoTil = Sys.Date(),
  enhetsUtvalg = 0, reshID = 0, traume = "")

tabAntOpphShMnd(RegData, datoTil = Sys.Date(), traume = "",
  antMnd = 12)

tabAntOpphSh5Aar(RegData, datoTil = Sys.Date(), traume = "")

tabAntOpphPasSh5Aar(RegData, gr = "opph", datoTil)

tabSkjemaTilknyttet(Data = AlleTab, moderSkjema = "Hoved",
  datoFra = "2017-01-01", datoTil = Sys.Date())

tabAntSkjema(Data = AlleTab, datoFra = "2017-01-01",
  datoTil = Sys.Date())

lagTabavFigAndeler(UtDataFraFig)

lagTabavFigGjsnGrVar(UtDataFraFig)

lagTabNevrKlass(HovedSkjema, datoFra = "2018-01-01",
  datoTil = Sys.Date())
}
\arguments{
\item{RegData}{data}

\item{datoTil}{sluttdato. Brukes i tabellene AntOpph per 12 mnd og Belegg}

\item{datoFra}{fra og med dato}

\item{Data}{Liste med alle datatabeller/skjema}

\item{moderSkjema}{Hvilket skjema man skal knytte oppfølgingene til}

\item{personIDvar}{Variabelen som angir pasientidentifikasjon}

\item{datoTil}{til og med dato
inheritParams NIRFigAndeler}
}
\description{
Fil som beregner div tabeller.Group of functions Description section

AntReg <- table(Data$HovedSkjema$ShNavn)
AntOppf <- cbind(Hoved = AntReg, 
                 apply(RaaTab[ ,-1], MARGIN=2, 
                       FUN=function(x) tapply(x,INDEX=RaaTab$Sykehus, sum))
                 )
addmargins(AntOppf, margin=1, FUN = list('Hele landet' = sum) )
AndelOppf <- (100*AntOppf / as.vector(AntReg))[,-1]
}
\details{
Detaljer. kommer senereGroup of functions Details paragraph.

Fil som inneholder funksjoner for å lage tabeller, i første rekke tellinger av personer 
Aktuelle tabeller:
-Belegg (samlerapport), antall opphold per år og enhet, ant. pasienter per år og enhet, ant opph. per måned og enhet.
-tabAntOpphSh12mnd: Antall opphold per måned og enhet siste 12 måneder fram til datoTil. 
RegData må inneholde InnDato.
-tabAntOpphSh5Aar:Antall opphold per år og enhet siste 5 år (inkl. inneværende år) fram til datoTil. 
RegData må inneholde Aar.
}
\section{Belegg (antall opphold, pasienter og rehabdøgn). Siste inntil 5 år eller siste inntil 12 måneder/kvartal/halvår}{
NA
}

\section{Liggetider. Liggetid tot.(OpphTot), Liggetid rehab.(DagerRehab), Liggetid før rehab (DagerTilRehab)}{
 
min/max, gjsn, median
}

\section{tabAntOpphShMnd antall opphold siste X (antMnd) mnd}{
NA
}

\section{Antall opphold siste 5 år}{
NA
}

\section{Antall registreringer/pasienter siste 5 år}{

Hmmm
}

\section{Tabell}{
 Antall og andel moder"skjema som har ulike typer registreringsskjema

 Antall av ulike typer skjema
}

\section{Vise figurdata som tabell, fordelingsfigur}{
NA
}

\section{Vise figurdata som tabell, sentralmål per sykshus}{
NA
}

\section{Nevrologisk klassifikasjon}{
NA
}

