% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NStabeller.R
\name{NordicScirtabeller}
\alias{NordicScirtabeller}
\alias{tabBelegg}
\alias{tabAntOpphShMnd}
\alias{tabAntOpphSh5Aar}
\alias{tabAntOpphPasSh5Aar}
\alias{tabSkjemaTilknyttetH}
\alias{tabAntSkjema}
\alias{lagTabavFig}
\title{Funksjoner for å lage tabeller Group of functions page title}
\usage{
tabBelegg(RegData, tidsenhet = "Aar", datoTil = Sys.Date(),
  enhetsUtvalg = 0, reshID = 0)

tabAntOpphShMnd(RegData, datoTil = Sys.Date(), antMnd = 12)

tabAntOpphSh5Aar(RegData, datoTil)

tabAntOpphPasSh5Aar(RegData, gr = "opph", datoTil)

tabSkjemaTilknyttetH(Data = AlleTab, datoFra = "2017-01-01",
  datoTil = Sys.Date())

tabAntSkjema(Data = AlleTab, datoFra = "2017-01-01", datoTil = Sys.Date())

lagTabavFig(UtDataFraFig)
}
\arguments{
\item{RegData}{data}

\item{datoTil}{sluttdato. Brukes i tabellene AntOpph per 12 mnd og Belegg}

\item{Data}{Liste med alle datatabeller/skjema}

\item{datoFra}{fra og med dato}

\item{personIDvar}{Variabelen som angir pasientidentifikasjon}

\item{datoTil}{til og med dato}

\item{Data}{Liste med alle datatabeller/skjema}

\item{datoFra}{fra og med dato}

\item{datoTil}{til og med dato}
}
\value{
Div tabeller
}
\description{
Fil som beregner div tabeller.Group of functions Description section
}
\details{
Detaljer. kommer senereGroup of functions Details paragraph.

Fil som inneholder funksjoner for å lage tabeller, i første rekke tellinger av personer 
Aktuelle tabeller:
-Belegg (samlerapport), antall opphold per år og enhet, ant. pasienter per år og enhet, ant opph. per måned og enhet.
-tabAntOpphSh12mnd: Antall opphold per måned og enhet siste 12 måneder fram til datoTil. 
RegData må inneholde InnDato.
-tabAntOpphSh5Aar:Antall opphold per år og enhet siste 5 år (inkl. inneværende år) fram til datoTil. 
RegData må inneholde Aar.
}
\section{Belegg (antall opphold, pasienter og intensivdøgn). Siste inntil 5 år eller siste inntil 12 måneder/kvartal/halvår}{
NA
}

\section{tabAntOpphShMnd antall opphold siste X (antMnd) mnd}{
NA
}

\section{Antall opphold siste 5 år}{
NA
}

\section{Antall registreringer/pasienter siste 5 år}{

Hmmm
}

\section{Tabell}{
 Antall og andel hovedskjema som har ulike typer registreringsskjema

 Antall av ulike typer skjema
}

\section{Vise figurdata som tabell}{
NA
}

