% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSFigAndeler.R
\name{NSFigAndeler}
\alias{NSFigAndeler}
\title{Søylediagram som viser fordeling av valgt variabel}
\usage{
NSFigAndeler(RegData, outfile = "", valgtVar, datoFra = "2010-01-01",
  datoTil = "2050-01-01", AIS = "", minald = 0, maxald = 130,
  erMann = 99, traume = "", enhetsUtvalg = 1, reshID, hentData = 0,
  preprosess = 1)
}
\arguments{
\item{RegData}{- ei dataramme med alle nødvendige variable fra registeret.}

\item{outfile}{- navn på fil figuren skrives ned til.}

\item{valgtVar}{- variabelen det skal genereres resultat for}

\item{datoFra}{<- '2010-01-01'    # min og max dato i utvalget vises alltid i figuren.}

\item{datoTil}{<- '2013-05-25'}

\item{AIS}{- AISgrad ved innleggelse alle(''), velge en eller flere fra A,B,C,D,E,U}

\item{minald}{- alder, fra og med}

\item{maxald}{- alder, til og med}

\item{erMann}{- kjønn, 1-menn, 0-kvinner, standard: '' (alt annet enn 0 og 1), dvs. begge}

\item{traume}{- 'ja','nei', standard: ikke valgt}

\item{enhetsUtvalg}{- 1:eget sykehus, 0:hele landet (standard)}

\item{reshID}{- avdelingsid for egen avdeling, standard: 0-hele landet.}

\item{hentData}{Gjør spørring mot database hvis data ikke er levert fra andre kilder.
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{preprosess}{Preprosesser data
0: Nei
1: Ja (Standard)}
}
\description{
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item Alder: Aldersfordeling, 15-årige grupper 
    \item AAis: A-E+U, innleggelse
    \item FAis: A-E+U, kontroll
    \item DagerRehab: Lengde på rehab.opphold (Utskrevet – Innl. rehab dato)
    \item DagerTilRehab: Tid fra skadedato til oppstart rehab [trauma/ikke], 5d interv
    \item OpphTot: Lengde på opphold – totalt (HosptlDy) 20d… >200, 
    \item Permisjon: Antall døgn ute av sykehus 
    \item UtTil: Hva pasienten skrives ut til. [PlaceDis]
    \item SkadeArsak: Årsak til skade [variabel: Scietiol]
    \item Pustehjelp[VentAssi].
   }
}
