% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSFigAndeler.R
\name{NSFigAndeler}
\alias{NSFigAndeler}
\title{Søylediagram som viser fordeling av valgt variabel}
\usage{
NSFigAndeler(
  RegData,
  outfile = "",
  valgtVar = "Alder",
  datoFra = "2010-01-01",
  datoTil = Sys.Date(),
  AIS = "",
  minald = 0,
  maxald = 130,
  erMann = 99,
  traume = "alle",
  nivaaUt = 99,
  enhetsUtvalg = 0,
  reshID = 0,
  hentData = 0,
  preprosess = 1,
  datoUt = 0,
  ...
)
}
\arguments{
\item{RegData}{- ei dataramme med alle nødvendige variable fra registeret.}

\item{outfile}{- navn på fil figuren skrives ned til.}

\item{valgtVar}{- variabelen det skal genereres resultat for}

\item{reshID}{- avdelingsid for egen avdeling, standard: 0-hele landet.}

\item{hentData}{Gjør spørring mot database hvis data ikke er levert fra andre kilder.
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{preprosess}{Preprosesser data
0: Nei
1: Ja (Standard)}

\item{datoUt}{Om man skal velge inn eller ut-dato som grunnlag for datofiltrering}

\item{figurtype}{angir hvilken figurtype som skal lages: andeler, gjsnGrVar}
}
\description{
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item Alder: Aldersfordeling, 15-årige grupper
    \item AAis: A-E+U, innleggelse
    \item FAis: A-E+U, kontroll
    \item DagerRehab: Lengde på rehab.opphold (Utskrevet – Innl. rehab dato)
    \item DagerTilRehab: Tid fra skadedato til oppstart rehab [trauma/ikke], 5d interv
    \item OpphTot: Lengde på opphold – totalt (HosptlDy) 20d… >200,
    \item Permisjon: Antall døgn ute av sykehus
    \item UtTil: Hva pasienten skrives ut til. [PlaceDis]
    \item SkadeArsak: Årsak til skade [variabel: Scietiol]
    \item Pustehjelp[VentAssi]
    \item LivsGen:  Fornøydhet med livet
    \item LivsFys:  Fornøydhet med fysisk helse
    \item LivsPsyk:  Fornøydhet med psykisk helse
    \item UrinInkontinens: Ufrivillig urinlekkasje
    \item UrinKirInngr: Kirurgiske inngrep i urinveiene
    \item UrinLegemidler: Bruk av legemidler som påvirker urinveiene
    \item UrinLegemidlerHvilke: Legemidler som påvirker urinveiene
    \item UrinTomBlareHoved: Blæretømming, hovedmetode
    \item UrinTomBlareTillegg: Blæretømming, tilleggsmetode
    \item TarmAvfmiddel: Bruk av perorale avføringsmidler
    \item TarmAvfmiddelHvilke: Perorale avføringsmidler
    \item TarmInkontinens: Hyppighet av fekal inkontinens
    \item TarmKirInngrep: Kirurgisk inngrep i mage/–tarm kanalen
    \item TarmKirInngrepHvilke: Kirurgiske inngrep i mage/–tarm kanalen
    \item TarmAvfHoved: Avføring, hovedmetode
    \item TarmAvfTillegg: Avføring, tilleggsmetode
   }
}
