% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSFigKvalIndIkkeNodvendigIkkeKlar.R
\name{NSFigKvalInd}
\alias{NSFigKvalInd}
\title{Søylediagram som viser flere kvalitetsindikatorer}
\usage{
NSFigKvalInd(NSdata, outfile = "", valgtVar, datoFra = "2010-01-01",
  datoTil = "2050-01-01", AIS = "", minald = 0, maxald = 130,
  erMann = 99, traume = "", paratetra = 99, enhetsUtvalg = 1,
  reshID, hentData = 0, preprosess = 1)
}
\arguments{
\item{NSdata}{- Alle nødvendige variable fra hvert skjema.}

\item{outfile}{- navn på fil figuren skrives ned til.}

\item{valgtVar}{- variabelen det skal genereres resultat for}

\item{reshID}{- avdelingsid for egen avdeling, standard: 0-hele landet.}

\item{hentData}{Gjør spørring mot database hvis data ikke er levert fra andre kilder.
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{preprosess}{Preprosesser data
0: Nei
1: Ja (Standard)}

\item{figurtype}{angir hvilken figurtype som skal lages: andeler, gjsnGrVar}
}
\value{
Søylediagram samling av kvalitetsindikatorer
}
\description{
Funksjonen genererer en figur med ei gruppe kvalitetsindikatorer.
}
\details{
Argumentet \emph{valgtVar} kan bare være "kvalInd".
}
