% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSRegDataSQL.R
\name{NSRegDataSQL}
\alias{NSRegDataSQL}
\title{Henter data for NordicScir fra  staging}
\usage{
NSRegDataSQL(valgtVar = "Alder")
}
\value{
RegData data frame
}
\description{
Spørring som henter og kobler sammen data fra ulike skjema i NordicScir
}
\details{
Aktuelle skjema: 
ActivityAndParticipationPerformanceFormDataContract: Aktivitetsfunksjon
ActivityAndParticipationSatisfactionFormDataContract: Aktivitetsfornøydhet
BowelFunctionFormDataContract: Tarmfunksjon (?)
ControlFormDataContract: Kontroll. Noen variable er oppfølging av målinger i hovedskjema
LifeQualityFormDataContract: Livskvalitet
MainFormDataContract: HOVEDSKJEMA
UrinaryTractFunctionFormDataContract: Urinfunksjon
}
