% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSRegDataSQL.R
\name{NSRegDataSQL}
\alias{NSRegDataSQL}
\title{Henter data for NorScir eller NordicScir fra database}
\usage{
NSRegDataSQL(
  valgtVar = "Alder",
  register = "norscir",
  koblSkjema = "Hoved",
  ...
)
}
\arguments{
\item{valgtVar}{Valg av variabel angir hvilket skjema som skal knyttes til hovedskjema
siden de fire første bokstavene i variabelnavnet utgjør et prefiks for
å identifisere hvilken tabell variabelen skal hentes fraaktuell tabell. Eks LivsAlder
Variabler uten prefiks hentes fra hovedtabellen (Main...)}

\item{register}{Hvilket register det skal hentes data for: 'norscir' (standard) eller 'nordicscir'}

\item{koblSkjema}{Hvilket skjema skal regnes som "hovedskjema". Standard: 'Hoved' (spinal_cord_injury_core_data_set_1),
'Kont' (control_form_2). Bare NorScir har kontrollskjema}
}
\value{
RegData data frame
}
\description{
Spørring som henter og kobler sammen data fra ulike skjema i valgt register
}
\details{
Aktuelle skjema:
spinal_cord_injury_core_data_set_1: HOVEDSKJEMA
bowel_function_5: Tarmfunksjon (?)
registration_of_quality_of_life_3: Livskvalitet
lower_urinary_tract_function_4: Urinfunksjon
activities_and_participation_performance_6: Aktivitetsfunksjon.
     Skjema kun i det norske registeret
activities_and_participation_satisfaction_7: Aktivitetsfornøydhet.
     Skjema kun i det norske registeret
control_form_2: Kontroll. Noen variable er oppfølging av målinger i hovedskjema.
     Skjema kun i det norske registeret
}
