% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSRegDataSQL.R
\name{NSRegDataSQL}
\alias{NSRegDataSQL}
\title{Henter data for NorScir eller NordicScir fra database}
\usage{
NSRegDataSQL(
  valgtVar = "Alder",
  register = "norscir",
  koblSkjema = "Hoved",
  ...
)
}
\arguments{
\item{valgtVar}{Valg av variabel angir hvilket skjema som skal knyttes til hovedskjema
siden de fire første bokstavene i variabelnavnet utgjør et prefiks for
å identifisere hvilken tabell variabelen skal hentes fraaktuell tabell. Eks LivsAlder
Variabler uten prefiks hentes fra hovedtabellen (Main...)}

\item{register}{Hvilket register det skal hentes data for: 'norscir' (standard) eller 'nordicscir'}

\item{koblSkjema}{Hvilket skjema skal regnes som "hovedskjema". Standard: 'Hoved' (mainformdatacontract),
'Kont' (controlformdatacontract). Bare NorScir har kontrollskjema}
}
\value{
RegData data frame
}
\description{
Spørring som henter og kobler sammen data fra ulike skjema i valgt register
}
\details{
Aktuelle skjema:
mainformdatacontract: HOVEDSKJEMA
bowelfunctionformdatacontract: Tarmfunksjon (?)
lifequalityformdatacontract: Livskvalitet
urinarytractfunctionformdatacontract: Urinfunksjon
ActivityAndParticipationPerformanceFormDataContract: Aktivitetsfunksjon.
     Skjema kun i det norske registeret
ActivityAndParticipationSatisfactionFormDataContract: Aktivitetsfornøydhet.
     Skjema kun i det norske registeret
controlformdatacontract: Kontroll. Noen variable er oppfølging av målinger i hovedskjema.
     Skjema kun i det norske registeret
}
