% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NStabeller.R
\name{tabAntOpphShMnd}
\alias{tabAntOpphShMnd}
\title{tabAntOpphShMnd antall opphold siste X (antMnd) mnd
-tabAntOpphSh12mnd: Antall opphold per måned og enhet siste 12 måneder fram til datoTil. 
-tabAntOpphSh5Aar:Antall opphold per år og enhet siste 5 år (inkl. inneværende år) fram til datoTil. 
RegData må inneholde Aar.}
\usage{
tabAntOpphShMnd(RegData, datoTil = Sys.Date(), traume = "", antMnd = 12)
}
\arguments{
\item{datoTil}{<- '2013-05-25'}

\item{traume}{- 'ja','nei', 'alle' standard: ikke valgt}
}
\description{
tabAntOpphShMnd antall opphold siste X (antMnd) mnd
-tabAntOpphSh12mnd: Antall opphold per måned og enhet siste 12 måneder fram til datoTil. 
-tabAntOpphSh5Aar:Antall opphold per år og enhet siste 5 år (inkl. inneværende år) fram til datoTil. 
RegData må inneholde Aar.
}
